backend stvari

+++ backend/src/main/java/timski/proekt/backend/Repository/ReviewRepository.java
@@ -16,7 +16,7 @@ public interface ReviewRepository extends JpaRepository<Review, Long> {
     List<Review> findReviewsByUser(User user);
     List<Review> findReviewsByCompany(Company company);
 
-    List<Review> findReviewsByRatingGreaterThan(int rating);
+    List<Review> findReviewsByRatingGreaterThan(Long rating);
 
     List<Review> findReviewsByTitle(String title);
 }
