+++ backend/src/main/java/timski/proekt/backend/Service/ReviewService.java
@@ -21,5 +21,5 @@ public interface ReviewService {
     List<Review> findByCompany(Company company);
 
     List<Review> findByUser(User user);
-    List<Review> findByRatingGreaterThan(int rating);
+    List<Review> findByRatingGreaterThan(long rating);
 }
